requirements.txt
feedgenerator
Jinja2
Pygments
PyYAML
